chore: deprecate and move nav bar value list item parse
